Improve the out-of-date warning message.
In case of self-hosted runner, the organization uses the
GitHub runner, but they also have their own self-hosted runner
"version", with specific configuration being improved over time
and through this increasing their own internal "runner version".

The way the message was written, sometimes results in
confusion, where developer and a devops person start mistakenly
assuming that the version is related to their own runner version,
and not the actual GitHub runner that runs on that hardware.
Time is then lost, until someone realizes that this has nothing
to do with the internal "runner version" itself, but with the
actual GitHub runner that needs to be upgraded.

Message was fixed to explicitely differentiate between
"GitHub runner" and the self-hosted runner itself to avoid
possible confusion and possibly save someone few minutes of time.